+++ src/components/Header/index.tsx
@@ -36,7 +36,7 @@ function Header() {
             href="https://github.com/nitishkgupta/seotoolsuite/blob/main/CHANGELOG.md"
             target="_blank"
             rel="nofollow"
-            className="block rounded-md border border-slate-200 px-2 py-1 text-sm font-medium text-black/60 transition duration-300 hover:bg-slate-100"
+            className="block rounded-md border border-slate-200 px-2 py-1 text-sm font-medium text-black/60 transition hover:bg-slate-100"
           >
             v{APP_VERSION}
           </Link>
@@ -46,29 +46,29 @@ function Header() {
             <Tooltip content="Keyword Research">
               <Link
                 href="/tool/keyword-research"
-                className={`flex items-center gap-1 px-2 py-2 text-sky-950 transition duration-300 hover:bg-slate-100 ${isToolActive("keyword-research") ? "bg-sky-950/10!" : ""}`}
+                className={`flex items-center gap-1 px-2 py-2 text-sky-950 transition hover:bg-slate-100 ${isToolActive("keyword-research") ? "bg-sky-950/10!" : ""}`}
               >
                 <TelescopeIcon size={24} />
               </Link>
             </Tooltip>
           </div>
           <DFSBalanceBox />
-          <div className="my-1 w-0.5 bg-slate-200"></div>
+          <div className="my-1 hidden w-0.5 bg-slate-200 lg:block"></div>
           <Tooltip content="Settings">
             <Link
               href="/account/settings"
-              className={`flex items-center gap-1 rounded-md border-2 border-slate-200 px-2 py-2 text-black/80 transition duration-300 hover:bg-slate-100 ${isSettingsPageActive() ? "bg-sky-950/10!" : ""}`}
+              className={`flex items-center gap-1 rounded-md border-2 border-slate-200 px-2 py-2 text-black/80 transition hover:bg-slate-100 ${isSettingsPageActive() ? "bg-sky-950/10!" : ""}`}
             >
               <SettingsIcon size={24} />
             </Link>
           </Tooltip>
-          <div className="my-1 w-0.5 bg-slate-200"></div>
+          <div className="my-1 hidden w-0.5 bg-slate-200 lg:block"></div>
           <Tooltip content="Feedback">
             <Link
               href="https://github.com/nitishkgupta/seotoolsuite/issues/new/choose"
               rel="nofollow"
               target="_blank"
-              className={`flex items-center gap-1 rounded-md border-2 border-slate-200 px-2 py-2 text-black/80 transition duration-300 hover:bg-slate-100`}
+              className={`flex items-center gap-1 rounded-md border-2 border-slate-200 px-2 py-2 text-black/80 transition hover:bg-slate-100`}
             >
               <SmilePlusIcon size={24} />
             </Link>