+++ src/services/DataForSEO/index.ts
@@ -1,4 +1,6 @@
 import axios from "axios";
+import UpstashRedis from "@/services/UpstashRedis";
+import { getLocalStorageItem } from "@/utils/localStorage";
 
 /**
  * DataForSEO service.
@@ -20,16 +22,50 @@ class DataForSEO {
   private PASSWORD: string;
 
   /**
+   * Sandbox mode.
+   */
+  private sandboxEnabled: boolean;
+
+  /**
+   * Enable caching.
+   */
+  private enableCaching: boolean;
+
+  /**
+   * Caching duration (in days).
+   */
+  private cachingDuration: number;
+
+  /**
+   * Upstash Redis instance.
+   */
+  private upstashRedis?: UpstashRedis;
+
+  /**
    * Constructor.
    * @param isSandbox - Whether to use the sandbox version of the API.
    */
-  constructor(username: string, password: string, isSandbox: boolean = false) {
+  constructor(
+    username: string,
+    password: string,
+    isSandbox: boolean = false,
+    enableCaching: boolean = false,
+  ) {
     this.API_BASE_URL = isSandbox
       ? "https://sandbox.dataforseo.com/v3"
       : "https://api.dataforseo.com/v3";
 
     this.USERNAME = username;
     this.PASSWORD = password;
+    this.sandboxEnabled = isSandbox;
+    this.enableCaching = enableCaching;
+    this.cachingDuration = Number(
+      getLocalStorageItem("CACHING_DURATION") ?? 30,
+    );
+
+    if (!this.sandboxEnabled && enableCaching) {
+      this.upstashRedis = new UpstashRedis();
+    }
   }
 
   /**
@@ -87,6 +123,21 @@ class DataForSEO {
     limit: number = 50,
     offset: number = 0,
   ) {
+    if (!this.sandboxEnabled && this.enableCaching && this.upstashRedis) {
+      let cachedData;
+      try {
+        cachedData = await this.upstashRedis.getData(
+          btoa(
+            `keyword-suggestions-${keyword}-${location_code}-${language_code}-${JSON.stringify(filters)}-${limit}-${offset}`,
+          ),
+        );
+      } catch (error) {
+        console.error(error);
+      }
+
+      if (cachedData) return JSON.parse(cachedData);
+    }
+
     try {
       const apiResponse = await axios.post(
         `${this.API_BASE_URL}/dataforseo_labs/google/keyword_suggestions/live`,
@@ -110,6 +161,27 @@ class DataForSEO {
         },
       );
 
+      const taskStatusCode = apiResponse?.data?.tasks[0]?.status_code ?? null;
+
+      if (
+        !this.sandboxEnabled &&
+        this.enableCaching &&
+        this.upstashRedis &&
+        taskStatusCode === 20000
+      ) {
+        try {
+          this.upstashRedis.setData(
+            btoa(
+              `keyword-suggestions-${keyword}-${location_code}-${language_code}-${JSON.stringify(filters)}-${limit}-${offset}`,
+            ),
+            JSON.stringify(apiResponse.data),
+            60 * 60 * 24 * this.cachingDuration,
+          );
+        } catch (error) {
+          console.error(error);
+        }
+      }
+
       return apiResponse.data;
     } catch (error) {
       throw error;
@@ -125,6 +197,21 @@ class DataForSEO {
     language_code: string = "en",
     includeClickstreamData: boolean = false,
   ) {
+    if (!this.sandboxEnabled && this.enableCaching && this.upstashRedis) {
+      let cachedData;
+      try {
+        cachedData = await this.upstashRedis.getData(
+          btoa(
+            `keywords-overview-${JSON.stringify(keywords)}-${location_code}-${language_code}-${includeClickstreamData}`,
+          ),
+        );
+      } catch (error) {
+        console.error(error);
+      }
+
+      if (cachedData) return JSON.parse(cachedData);
+    }
+
     try {
       const apiResponse = await axios.post(
         `${this.API_BASE_URL}/dataforseo_labs/google/keyword_overview/live`,
@@ -144,6 +231,27 @@ class DataForSEO {
           },
         },
       );
+
+      const taskStatusCode = apiResponse?.data?.tasks[0]?.status_code ?? null;
+
+      if (
+        !this.sandboxEnabled &&
+        this.enableCaching &&
+        this.upstashRedis &&
+        taskStatusCode === 20000
+      ) {
+        try {
+          this.upstashRedis.setData(
+            btoa(
+              `keywords-overview-${JSON.stringify(keywords)}-${location_code}-${language_code}-${includeClickstreamData}`,
+            ),
+            JSON.stringify(apiResponse.data),
+            60 * 60 * 24 * this.cachingDuration,
+          );
+        } catch (error) {
+          console.error(error);
+        }
+      }
 
       return apiResponse.data;
     } catch (error) {